refactor: update Next.js configuration and remove deprecated middleware
- Added new rewrites for marketing static paths and homepage in `next.config.ts`.
- Removed obsolete middleware and related components from the project structure, including layout and feature sections.
- Introduced a new footer component to maintain site navigation and branding.
