+++ docs/src/app/docs/footer.tsx
@@ -1,7 +1,44 @@
 import Link from 'next/link';
 
 import { C15TLogo } from '~/components/logo';
-import { siteConfig } from '../config';
+
+export const siteConfig = {
+	footer: {
+		links: [
+			{
+				title: 'Product',
+				items: [
+					{ text: 'Documentation', url: '/docs' },
+					{ text: 'Components', url: '/docs/components/react/cookie-banner' },
+				],
+			},
+			{
+				title: 'Company',
+				items: [
+					{
+						text: 'GitHub',
+						url: 'https://github.com/c15t/c15t',
+						external: true,
+					},
+					{
+						text: 'Contact',
+						url: 'mailto:support@c15t.com',
+						external: true,
+					},
+				],
+			},
+			{
+				title: 'Legal',
+				items: [
+					{ text: 'Privacy Policy', url: '/docs/legals/privacy-policy' },
+					{ text: 'Cookie Policy', url: '/docs/legals/cookie-policy' },
+				],
+			},
+		],
+		bottomText:
+			'Leverage native React components for seamless integration and high performance in a robust Consent Management solution that empowers your development team while prioritizing privacy and compliance.',
+	},
+};
 
 export function Footer() {
 	return (
