+++ docs/src/app/docs/layout.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react';
 import { DocsLayout } from '~/components/layouts/notebook';
 import { source } from '~/lib/source';
-import { Footer } from '../(home)/_components/footer';
+import { Footer } from './footer';
 import { docsOptions } from './layout.config';
 
 export default function Layout({ children }: { children: ReactNode }) {
